Separate content-tests
